requirements.txt
mkdocs==1.2.1
mkdocs-material==7.1.8
mkdocs_latest_release_plugin==0.1.2
markdown-full-yaml-metadata==2.0.1
python-frontmatter==1.0.0
yamale==3.0.7
pymdown-extensions==8.2
mkdocs-git-authors-plugin==0.4.1
mkdocs-git-revision-date-localized-plugin==0.9.2
